Extract full v2 driver to common
